[Inhance] Used TypeScript - Haven't done for event Type and etc
