--- src/components/InhancedModal/OptionalModal.tsx
@@ -1,18 +1,0 @@
-import * as React from 'react';
-import { useState } from 'react';
-import {View, Text, TouchableOpacity} from 'react-native'
-import styled, { css } from "styled-components/native";
-
-function OptionalModal (){
-    return(
-        <View>
-            <StyledOption>
-                
-            </StyledOption>
-
-            
-        </View>
-    )
-}
-
-export default OptionalModal;
