--- src/components/Modal/Button.tsx
@@ -1,17 +1,0 @@
-import React, {useState} from 'react';
-import styled, { css } from "styled-components/native";
-
-function Button() {
-
-  return (
-      <TriggerBtn> "Open modal"</TriggerBtn>
-  );
-}
-
-export default Button;
-
-const TriggerBtn = styled.Button`
-  width: 100px;
-  height: 30px;
-  z-index: 9;
-`;
