+++ src/components/Modal/CloseButton.tsx
