+++ src/components/Modal/Modal.tsx
@@ -1,92 +1,141 @@
-import React, {useState} from 'react';
-import {View, ViewStyle, TextStyle, Text, TouchableOpacity} from 'react-native'
-import styled, { css } from "styled-components/native";
-import BackDrop from './BackDrop'
+import * as React from 'react';
+import { useRef, useState } from 'react';
+import { Animated, ViewStyle, GestureResponderEvent } from 'react-native';
+import styled from 'styled-components/native';
+import CloseButton from './CloseButton';
 
 interface Props {
-  containerStyle?: ViewStyle;
-  disabledStyle?: ViewStyle;
-  style?: ViewStyle;
-  textStyle?: TextStyle;
-  open?: boolean;
-  setOpen?: () => void;
+    visible?: boolean | undefined;
+    closable?: boolean | undefined;
+    backDropClosable?: boolean | undefined;
+    onClose?: () => void | undefined;
+    children: React.ReactElement; 
+    style?: ViewStyle | undefined;
+    isAnimated?: boolean | undefined;
+    animationType?: string | undefined;
+    animationSpeed?: number | undefined;
 }
 
-const StyledModal = styled.View`
-  background-color: #ffff;
-  border-radius: 10;
-  align-self: center;
-  width: 300px;
-  height: 300px;
-  border: 2px solid #088EDF;
-  box-shadow: rgba(0,0,0,.25) 0 10px 20px;
-  align-items: center;
-  justify-content: center;
-  opacity: 1;
-`;
-
-const StyledDisable = styled(StyledModal)`
-  height: 0;
-  opacity: 0;
-`;     
-
-const StyledButton = styled.TouchableOpacity`
- background-color: #119CF2;
- width: 100px;
- height: 30px;
- border-radius: 5;
- justify-content: center;
- box-shadow: rgba(0,0,0,.15) 0 10px 20px;
-`;
-
-const BtnText = styled.Text` 
- color: #ffff;
- font-weight: 800;
- text-align: center;
- `
-
- 
-
-function Modal ( props : Props ):React.ReactElement {
-
-   const {
-     containerStyle,
-     style,
-     textStyle,
-     open,
-     setOpen,
+function Modal(props : Props) {
+    const {
+      visible,
+      closable,
+      backDropClosable,
+      onClose,
+      children,
+      style,
+      isAnimated,
+      animationType, 
+      animationSpeed 
     } = props
 
-  if(open === false){
-    return(
-      <StyledDisable style={[containerStyle]}>
-      </StyledDisable>
+    const onbackDropPress = () =>  {
+      if (e.target === e.currentTarget) {
+        onClose(e)
+      }
+    }
+  
+    const close = () => {
+      if (onClose) {
+        onClose(e)
+      }
+    }
+
+    const fadeAnimation = useRef(new Animated.Value(0)).current;
+    const defaultAnimation = useRef(new Animated.Value(500)).current;
+    const zoomAnimation =  useRef(new Animated.Value(0.8)).current; 
+    // const [type, setType] = useState("")
+
+
+    if (animationType === "default"){
+        Animated.timing(defaultAnimation, {
+            toValue: -50,
+            duration: animationSpeed,
+            useNativeDriver: true
+          }).start();
+    }  
+ 
+    if (animationType === "fadeIn"){
+        Animated.timing(fadeAnimation, {
+          toValue: 1,
+          duration: animationSpeed,
+          useNativeDriver: true
+        }).start();
+      };
+
+    if (animationType === "zoom"){
+        Animated.timing(zoomAnimation, {
+            toValue: 1,
+            duration: animationSpeed,
+            useNativeDriver: true
+          }).start();
+    }
+
+    return (
+      <>
+        <ModalOverlay visible={visible} onPress={backDropClosable? onbackDropPress : null} />
+        <ModalWrapper
+          as={Animated.View} 
+          visible={visible}
+          style={
+              [ 
+                animationType === "default" ? 
+                {transform : [{translateY : defaultAnimation}]} : null,
+                animationType === "fadeIn" ? 
+                {opacity: fadeAnimation} : null,
+                animationType === "zoom" ? 
+                {transform : [{scale: zoomAnimation}]} : null
+         ]}
+        >
+          <ModalInner style={style}>
+            { closable && 
+             <CloseButton handlePress={close} />
+                }
+              {children}  
+          </ModalInner>
+        </ModalWrapper>
+      </>
     )
-  } 
-
-  console.log("modal props", props )
-
-  return (
-    <View style={containerStyle}>
-      <StyledModal style={style}> 
-        <Text style={textStyle}>안녕? 난 모달</Text>
-        <StyledButton onPress={() => {setOpen(false)}}>
-          <BtnText>닫기</BtnText>
-        </StyledButton>
-      </StyledModal>
-    </View>
-  )
-}
-
-  export default Modal;
+  }
 
 
+
+const ModalWrapper = styled.View`
+  opacity: ${(props) => (props.visible ? 1  : 0)};
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  z-index: 1000;
+`
+
+const ModalOverlay = styled.TouchableOpacity`
+    z-index: -1;
+    position: absolute;
+    right: 0;
+    bottom: 0;
+    top: 0;
+    left: 0;
+    background-color: rgba(0,0,0,0.7);
+`
+
+const ModalInner = styled.View`
+  box-shadow: rgba(0,0,0,.25) 0 10px 20px;
+  background-color: #fff;
+  border-radius: 10px;
+  width: 360px;
+  margin: 0 auto;
+  padding: 40px 20px;
+  top: -30%;
+`
 
 
 Modal.defaultProps = {
-    open: true
-}
+    closeBtnVisible: true,
+    closable: true,
+    visible: false,
+    animationType: "default",
+    animationSpeed: 1000
+  }
 
-
-
-
+export default Modal;
